Eliminated some errors from the WarlockText
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@622 bb130017-240f-0410-9613-9c2c827820fb

+++ core/cc.warlock.rcp/src/cc/warlock/rcp/ui/WarlockText.java
@@ -576,7 +576,7 @@ public class WarlockText implements LineBackgroundListener, Drawable {
 		textWidget.internal_dispose_GC(handle, data);
 	}
 
-	public int internal_new_GC(GCData data) {
+	public long internal_new_GC(GCData data) {
 		return textWidget.internal_new_GC(data);
 	}
 }
